all: rename AdserverAdtarget to adtrgtme
