+++ adapters/adtrgtme/adtrgtme.go
@@ -1,4 +1,4 @@
-package adserver_adtarget
+package adtrgtme
 
 import (
 	"encoding/json"
@@ -17,7 +17,7 @@ type adapter struct {
 	endpoint string
 }
 
-// Builder builds a new instance of the AdserverAdtarget adapter for the given bidder with the given config.
+// Builder builds a new instance of the Adtrgtme adapter for the given bidder with the given config.
 func Builder(bidderName openrtb_ext.BidderName, config config.Adapter) (adapters.Bidder, error) {
 	bidder := &adapter{
 		endpoint: config.Endpoint,
