+++ adapters/adtrgtme/adtrgtmetest/exemplary/banner-app.json
