+++ adapters/adtrgtme/adtrgtmetest/supplemental/empty-seatbid-array.json
