+++ adapters/adtrgtme/adtrgtmetest/supplemental/not-provided-site-or-app.json
