+++ adapters/adtrgtme/adtrgtmetest/supplemental/status-code-bad-request.json
