+++ adapters/adtrgtme/adtrgtmetest/supplemental/unsupported-bid-type-native.json
