+++ adapters/adtrgtme/adtrgtmetest/supplemental/unsupported-bid-type-video.json
