--- openrtb_ext/imp_adtargetme.go
@@ -1,4 +1,0 @@
-package openrtb_ext
-
-// ExtImpAdserverAdtarget defines the contract for bidrequest.imp[i].ext.adserver_adtarget
-type ExtImpAdserverAdtarget struct{}
