+++ openrtb_ext/imp_adtrgtme.go
@@ -1,0 +1,4 @@
+package openrtb_ext
+
+// ExtImpAdtrgtme defines the contract for bidrequest.imp[i].ext.adtrgtme
+type ExtImpAdtrgtme struct{}
