--- static/bidder-params/adserver_adtarget.json
@@ -1,8 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-04/schema#",
-  "title": "AdserverAdtarget Adapter Params",
-  "description": "A schema which validates params accepted by the AdserverAdtarget adapter",
-  "type": "object",
-
-  "properties": {}
-}
