+++ static/bidder-params/adtrgtme.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "title": "Adtrgtme adapter params",
+  "description": "A schema which validates params accepted by the Adtrgtme adapter",
+  "type": "object",
+
+  "properties": {}
+}
